@@ -57,49 +66,48 @@
     "babel-preset-react": "^6.5.0",
     "babel-preset-stage-1": "^6.3.13",
     "babel-register": "^6.4.3",
-    "eslint": "^1.10.3",
-    "eslint-config-rackt": "^1.1.1",
-    "eslint-plugin-react": "^3.15.0",
+    "eslint": "^3.15.0",
+    "eslint-config-react-app": "^0.5.0",
+    "eslint-plugin-flowtype": "^2.29.2",
+    "eslint-plugin-import": "^2.2.0",
+    "eslint-plugin-jsx-a11y": "^4.0.0",
+    "eslint-plugin-react": "^6.8.0",
     "expect": "^1.13.0",
     "history": "^3.0.0",
     "isparta": "^4.0.0",
     "isparta-loader": "^2.0.0",
-    "karma": "^0.13.3",
-    "karma-chrome-launcher": "^0.2.0",
-    "karma-coverage": "^0.5.3",
-    "karma-firefox-launcher": "^0.1.7",
-    "karma-ie-launcher": "^0.2.0",
-    "karma-mocha": "^0.2.0",
-    "karma-mocha-reporter": "^1.0.4",
-    "karma-safari-launcher": "^0.1.1",
+    "karma": "^1.4.1",
+    "karma-coverage": "^1.1.1",
+    "karma-firefox-launcher": "^1.0.0",
+    "karma-mocha": "^1.3.0",
+    "karma-mocha-reporter": "^2.2.2",
     "karma-sourcemap-loader": "^0.3.5",
-    "karma-webpack": "^1.7.0",
-    "mocha": "^2.3.4",
-    "react": "^0.14.3",
-    "react-dom": "^0.14.3",
-    "react-redux": "^4.4.0",
+    "karma-webpack": "^2.0.2",
+    "mocha": "^3.2.0",
+    "react": "^15.4.2",
+    "react-dom": "^15.4.2",
+    "react-redux": "^5.0.2",
     "react-router": "^3.0.0",
     "redux": "^3.0.4",
     "redux-devtools": "^3.0.0",
     "redux-devtools-dock-monitor": "^1.0.1",
     "redux-devtools-log-monitor": "^1.0.1",
-    "webpack": "^1.12.9"
+    "webpack": "^2.2.1"
   },
   "directories": {},
   "dist": {
-    "shasum": "9b1fde4e70106c50f47214e12bdd888cfb96e2a6",
-    "tarball": "https://registry.npmjs.org/react-router-redux/-/react-router-redux-4.0.7.tgz"
+    "shasum": "227403596b5151e182377dab835b5d45f0f8054e",
+    "tarball": "https://registry.npmjs.org/react-router-redux/-/react-router-redux-4.0.8.tgz"
   },
   "files": [
     "*.md",
-    "LICENSE",
     "dist",
+    "LICENSE",
     "lib",
     "src"
   ],
-  "gitHead": "3dfbddef27cbe4435ed96e63fcc05dc63081e001",
+  "gitHead": "a2825ac0b8e34662624455ab98e29e2eb27f8096",
   "homepage": "https://github.com/reactjs/react-router-redux#readme",
-  "installable": true,
   "keywords": [
     "react",
     "redux",
