@@ -1,33 +1,42 @@
 {
   "_args": [
     [
-      "webpack-dev-middleware@^1.4.0",
+      {
+        "raw": "webpack-dev-middleware@^1.4.0",
+        "scope": null,
+        "escapedName": "webpack-dev-middleware",
+        "name": "webpack-dev-middleware",
+        "rawSpec": "^1.4.0",
+        "spec": ">=1.4.0 <2.0.0",
+        "type": "range"
+      },
       "/Users/conor.britain/Documents/Projects/Other/Fantasy Baseball/auction-tool-react"
     ]
   ],
   "_from": "webpack-dev-middleware@>=1.4.0 <2.0.0",
-  "_id": "webpack-dev-middleware@1.9.0",
+  "_id": "webpack-dev-middleware@1.10.1",
   "_inCache": true,
   "_location": "/webpack-dev-middleware",
-  "_nodeVersion": "6.8.1",
+  "_nodeVersion": "6.9.5",
   "_npmOperationalInternal": {
-    "host": "packages-18-east.internal.npmjs.com",
-    "tmp": "tmp/webpack-dev-middleware-1.9.0.tgz_1481831937321_0.807177921757102"
+    "host": "packages-12-west.internal.npmjs.com",
+    "tmp": "tmp/webpack-dev-middleware-1.10.1.tgz_1487426124354_0.9857836843002588"
   },
   "_npmUser": {
-    "email": "kees@webduck.nl",
-    "name": "spacek33z"
+    "name": "spacek33z",
+    "email": "kees@webduck.nl"
   },
-  "_npmVersion": "3.10.8",
+  "_npmVersion": "3.10.10",
   "_phantomChildren": {
     "errno": "0.1.4",
-    "readable-stream": "2.2.2"
+    "readable-stream": "2.2.6"
   },
   "_requested": {
-    "name": "webpack-dev-middleware",
     "raw": "webpack-dev-middleware@^1.4.0",
-    "rawSpec": "^1.4.0",
     "scope": null,
+    "escapedName": "webpack-dev-middleware",
+    "name": "webpack-dev-middleware",
+    "rawSpec": "^1.4.0",
     "spec": ">=1.4.0 <2.0.0",
     "type": "range"
   },
@@ -35,8 +44,8 @@
     "#DEV:/",
     "/webpack-dev-server"
   ],
-  "_resolved": "https://registry.npmjs.org/webpack-dev-middleware/-/webpack-dev-middleware-1.9.0.tgz",
-  "_shasum": "a1c67a3dfd8a5c5d62740aa0babe61758b4c84aa",
+  "_resolved": "https://registry.npmjs.org/webpack-dev-middleware/-/webpack-dev-middleware-1.10.1.tgz",
+  "_shasum": "c6b4cf428139cf1aefbe06a0c00fdb4f8da2f893",
   "_shrinkwrap": null,
   "_spec": "webpack-dev-middleware@^1.4.0",
   "_where": "/Users/conor.britain/Documents/Projects/Other/Fantasy Baseball/auction-tool-react",
@@ -64,23 +73,22 @@
     "should": "^11.1.0",
     "sinon": "^1.17.5",
     "supertest": "^2.0.0",
-    "webpack": "^2.1.0-beta.22"
+    "webpack": "^2.2.0"
   },
   "directories": {},
   "dist": {
-    "shasum": "a1c67a3dfd8a5c5d62740aa0babe61758b4c84aa",
-    "tarball": "https://registry.npmjs.org/webpack-dev-middleware/-/webpack-dev-middleware-1.9.0.tgz"
+    "shasum": "c6b4cf428139cf1aefbe06a0c00fdb4f8da2f893",
+    "tarball": "https://registry.npmjs.org/webpack-dev-middleware/-/webpack-dev-middleware-1.10.1.tgz"
   },
   "engines": {
     "node": ">=0.6"
   },
   "files": [
-    "lib/",
-    "middleware.js"
+    "middleware.js",
+    "lib/"
   ],
-  "gitHead": "fcd8c29ea3e613669c8b59519be2f639daabbdaf",
+  "gitHead": "8e870fdb313e83bf2562e4fd4d5735035074903f",
   "homepage": "http://github.com/webpack/webpack-dev-middleware",
-  "installable": true,
   "license": "MIT",
   "main": "middleware.js",
   "maintainers": [
@@ -106,6 +114,7 @@
   "peerDependencies": {
     "webpack": "1 || ^2.1.0-beta || ^2.2.0-rc.0"
   },
+  "readme": "ERROR: No README data found!",
   "repository": {
     "type": "git",
     "url": "git+https://github.com/webpack/webpack-dev-middleware.git"
@@ -118,5 +127,5 @@
     "test": "mocha --full-trace --check-leaks",
     "travis": "npm run cover -- --report lcovonly && npm run lint"
   },
-  "version": "1.9.0"
+  "version": "1.10.1"
 }
